refactor(cli): organize command and core sources
