+++ apps/cli/src/__tests__/build-info.test.ts
@@ -2,7 +2,7 @@ import { describe, expect, it } from "vitest";
 import packageManifest from "../../package.json" with { type: "json" };
 import { CHANNEL, CLI_PACKAGE_NAME, CLI_VERSION } from "../build-info.js";
 import { createProgram } from "../cli/program.js";
-import { channelConfig } from "../core/channel.js";
+import { channelConfig } from "../core/channel/config.js";
 
 describe("CLI build information", () => {
   it("derives the source identity from the package manifest", () => {
