+++ apps/cli/src/__tests__/daemon-owner.test.ts
@@ -2,7 +2,7 @@ import { mkdtemp, rm, writeFile } from "node:fs/promises";
 import { tmpdir } from "node:os";
 import { join } from "node:path";
 import { afterEach, describe, expect, it } from "vitest";
-import { acquireDaemonOwner } from "../core/daemon-owner.js";
+import { acquireDaemonOwner } from "../core/daemon/ownership.js";
 
 const directories: string[] = [];
 afterEach(async () => Promise.all(directories.splice(0).map((path) => rm(path, { recursive: true, force: true }))));
