+++ apps/cli/src/cli/index.ts
@@ -1,6 +1,6 @@
 #!/usr/bin/env node
 
-import "../core/channel-env.js";
+import "../core/channel/environment.js";
 import { createProgram } from "./program.js";
 
 await createProgram().parseAsync(process.argv);
