+++ apps/cli/src/commands/computer/list.ts
@@ -1,0 +1,12 @@
+import type { Command } from "commander";
+import { formatComputerList } from "../../core/computer/formatting.js";
+import { listComputers } from "../../core/computer/queries.js";
+
+export function registerComputerListCommand(computer: Command): void {
+  computer
+    .command("list")
+    .description("List Computers")
+    .action(async () => {
+      process.stdout.write(`${formatComputerList(await listComputers())}\n`);
+    });
+}
