+++ apps/cli/src/core/agent/context.ts
