+++ apps/cli/src/core/channel/config.ts
@@ -1,4 +1,4 @@
 import { getChannelConfig } from "@opentag/shared";
-import { CHANNEL } from "../build-info.js";
+import { CHANNEL } from "../../build-info.js";
 
 export const channelConfig = getChannelConfig(CHANNEL);
