+++ apps/cli/src/core/channel/environment.ts
@@ -1,4 +1,4 @@
-import { channelConfig } from "./channel.js";
+import { channelConfig } from "./config.js";
 
 if (!process.env.OPENTAG_HOME) {
   process.env.OPENTAG_HOME = channelConfig.defaultHome;
