+++ apps/cli/src/core/selection/team.ts
